+++ certbot/docs/api/certbot.ocsp.rst
@@ -1,0 +1,7 @@
+certbot.ocsp package
+======================
+
+.. automodule:: certbot.ocsp
+    :members:
+    :undoc-members:
+    :show-inheritance:
